refactor(upstream): service/repository interfaces
- adds PersistenceException
