+++ src/main/java/com/cbfacademy/restapiexercise/core/PersistenceException.java
@@ -1,0 +1,7 @@
+package com.cbfacademy.restapiexercise.core;
+
+public class PersistenceException extends RuntimeException {
+    public PersistenceException(String message) {
+        super(message);
+    }
+}
