+++ src/main/java/com/cbfacademy/restapiexercise/core/Repository.java
@@ -1,4 +1,4 @@
-package com.cbfacademy.restapiexercise;
+package com.cbfacademy.restapiexercise.core;
 
 import java.io.Serializable;
 import java.util.List;
@@ -10,7 +10,7 @@ public interface Repository<T, ID extends Serializable> {
      *
      * @return a list of all entities
      */
-    List<T> retrieveAll();
+    List<T> retrieveAll() throws PersistenceException;
 
     /**
      * Finds an entity by its unique identifier.
@@ -18,7 +18,7 @@ public interface Repository<T, ID extends Serializable> {
      * @param id the identifier of the entity
      * @return the found entity, or null if no such entity exists
      */
-    T retrieve(ID id);
+    T retrieve(ID id) throws IllegalArgumentException, PersistenceException;
 
     /**
      * Creates a new entity in the repository.
@@ -26,15 +26,15 @@ public interface Repository<T, ID extends Serializable> {
      * @param entity the {@code <T>} to create
      * @return the created entity
      */
-    T create(T entity);
+    T create(T entity) throws IllegalArgumentException, PersistenceException;
 
     /**
      * Deletes an entity from the repository based on its unique identifier.
      *
-     * @param id the identifier of the entity to be deleted
+     * @param entity the entity to update
      * @return true if the entity was successfully deleted; otherwise false
      */
-    Boolean delete(ID id);
+    void delete(T entity) throws IllegalArgumentException, PersistenceException;
 
     /**
      * Updates an existing entity in the repository.
@@ -42,6 +42,6 @@ public interface Repository<T, ID extends Serializable> {
      * @param entity the entity to update
      * @return the updated entity
      */
-    T update(T entity);
+    T update(T entity) throws IllegalArgumentException, PersistenceException;
 
 }
